# docs/swagger.yaml
basePath: /api
definitions:
...
  models.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      component:
        type: string
      created_at:
        description: 时间戳
        type: string
      external:
        type: boolean
      icon:
        type: string
      id:
        type: integer
      keep_alive:
        type: boolean
      meta:
        description: Meta 信息 (JSON)
        type: string
      name:
        type: string
      parent:
        allOf:
        - $ref: '#/definitions/models.Menu'
        description: 关联
      parent_id:
        type: integer
      path:
        type: string
      permission:
        description: 权限相关
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      sort:
        type: integer
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      visible:
        type: integer
    type: object
...
  models.Role:
    properties:
      code:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menus:
        items:
          $ref: '#/definitions/models.Menu'
        type: array
      name:
        type: string
      perm_list:
        items:
          type: string
        type: array
      status:
        type: integer
      updated_at:
        type: string
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.User:
    properties:
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_super_admin:
        description: Virtual field, not stored in database
        type: boolean
      nickname:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  response.PagedList:
    properties:
      items: {}
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
    type: object
  response.PaginationMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: Business status code
        type: integer
      data:
        description: Response data
      message:
        description: Response message
        type: string
      trace_id:
        description: Trace ID for request tracking
        type: string
    type: object
...
  services.CreateUserRequest:
    properties:
      avatar:
        type: string
      email:
        type: string
      nickname:
        type: string
      password:
        type: string
      role_ids:
        items:
          type: integer
        type: array
      status:
        type: integer
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A modern Go admin scaffold API server.
  title: Go Admin Scaffold API
  version: "1.0"
paths:
...
  /admin/v1/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of users with optional search filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Username filter
        in: query
        name: username
        type: string
      - description: Email filter
        in: query
        name: email
        type: string
      - description: Status filter (0=inactive, 1=active)
        in: query
        name: status
        type: integer
      - description: Role ID filter
        in: query
        name: role_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PagedList'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: List users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/services.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create user
      tags:
      - users
  /admin/v1/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get user
      tags:
      - users
...
swagger: "2.0"